analyzetestgroups: handle submission directories

+++ examples/arithmetic2/submissions/partially_accepted/arithmetic_d.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main() {
-	long double a, b, c;
+	double a, b, c;
 	cin >> a >> b >> c;
 	cout << setprecision(18) << fixed << a * b / c << endl;
 }
