+++ examples/arithmetic2/submissions/partially_accepted/arithmetic_dir/main.cpp
@@ -1,8 +1,9 @@
 #include <bits/stdc++.h>
+// @EXPECTED_GRADES@ AC AC WA WA
 using namespace std;
 
 int main() {
-	long double a, b, c;
+	double a, b, c;
 	cin >> a >> b >> c;
 	cout << setprecision(18) << fixed << a * b / c << endl;
 }
